+++ resources/recipes/default.rb
@@ -1,12 +1,9 @@
-#
-# Cookbook Name:: n2klocd
+# Cookbook:: n2klocd
 # Recipe:: default
-#
-# redborder
-#
-#
+# Copyright:: 2024, redborder
+# License:: Affero General Public License, Version 3
 
-n2klocd_config "config" do
-  name node["hostname"]
+n2klocd_config 'config' do
+  name node['hostname']
   action :add
 end