+++ CHANGELOG.md
@@ -1,5 +1,17 @@
-n2klocd CHANGELOG
+cookbook-n2klocd CHANGELOG
 ===============
+
+## 0.0.3
+
+  - Miguel Negron
+    - [61f3519] lint resources
+    - [369ffa2] lint recipes
+    - [f7e8cb1] lint providers
+    - [d3f625f] lint libraries
+    - [a211844] lint attributes
+    - [f0ece71] Update README.md
+    - [8719ba2] Update rpm.yml
+    - [f5165e2] Update metadata.rb
 
 0.0.1
 -----